requirements.txt
gpiozero==1.5.1
requests==2.22.0
python-dotenv==0.11.0
